+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## 0.0.10 / 2024-03-xx
 **New**
 - Add `Store.getState` method to directly query state.
-- Cache previous full state value for `Store.getState`.
+- Cache previous value for full retrieval.
 - Remove `Selector.removeAll`.
 
 **Changed**